extract services and middlewares into separate file to avoid circular dependencies
