+++ api/src/routers/authentication/index.ts
@@ -1,15 +1,10 @@
 import { Router } from 'express';
-import { CONFIG } from '../../config';
 import { ProveOwnershipPostBodySchema } from '../../models/schemas/request-response-body/authentication-bodies';
 import { AuthenticationRoutes } from '../../routes/authentication';
-import { AuthenticationService } from '../../services/authentication-service';
 import { Logger } from '../../utils/logger';
-import { apiKeyMiddleware, ssiService, validate } from '../helper';
-import { userService } from '../identity';
+import { authenticationService } from '../services';
+import { apiKeyMiddleware, validate } from '../middlewares';
 
-const { serverSecret, jwtExpiration } = CONFIG;
-
-export const authenticationService = new AuthenticationService(userService, ssiService, { jwtExpiration, serverSecret });
 const authenticationRoutes = new AuthenticationRoutes(authenticationService, Logger.getInstance());
 const { getNonce, proveOwnership } = authenticationRoutes;
