+++ api/src/routers/channel-info/index.ts
@@ -1,12 +1,10 @@
 import { Router } from 'express';
 import { ChannelInfoSchema } from '../../models/schemas/channel-info';
 import { ChannelInfoRoutes } from '../../routes/channel-info';
-import { ChannelInfoService } from '../../services/channel-info-service';
 import { Logger } from '../../utils/logger';
-import { apiKeyMiddleware, authMiddleWare, authorizationService, validate } from '../helper';
-import { userService } from '../identity';
+import { authorizationService, channelInfoService } from '../services';
+import { apiKeyMiddleware, authMiddleWare, validate } from '../middlewares';
 
-export const channelInfoService = new ChannelInfoService(userService);
 const channelInfoRoutes = new ChannelInfoRoutes(channelInfoService, authorizationService, Logger.getInstance());
 const { getChannelInfo, addChannelInfo, updateChannelInfo, deleteChannelInfo, searchChannelInfo } = channelInfoRoutes;
