+++ api/src/routers/channel/index.ts
@@ -1,15 +1,10 @@
 import { Router } from 'express';
-import { CONFIG } from '../../config';
 import { AddChannelLogBodySchema, CreateChannelBodySchema } from '../../models/schemas/request-response-body/channel-bodies';
 import { ChannelRoutes } from '../../routes/channel';
-import { ChannelService } from '../../services/channel-service';
 import { Logger } from '../../utils/logger';
-import { channelInfoService } from '../channel-info';
-import { apiKeyMiddleware, authMiddleWare, validate } from '../helper';
-import { streamsService, subscriptionPool, subscriptionService } from '../subscription';
+import { channelService } from '../services';
+import { apiKeyMiddleware, authMiddleWare, validate } from '../middlewares';
 
-const config = CONFIG.streamsConfig;
-export const channelService = new ChannelService(streamsService, channelInfoService, subscriptionService, subscriptionPool, config);
 const channelRoutes = new ChannelRoutes(channelService, Logger.getInstance());
 const { addLogs, createChannel, getLogs } = channelRoutes;
